refactor: move tests from Internal to normal test project
