+++ Tests/Mockolate.Tests/Exceptions/MockVerificationTimeoutExceptionTests.cs
@@ -1,7 +1,7 @@
 using aweXpect.Chronology;
 using Mockolate.Exceptions;
 
-namespace Mockolate.Internal.Tests.Exceptions;
+namespace Mockolate.Tests.Exceptions;
 
 public class MockVerificationTimeoutExceptionTests
 {
