+++ Tests/Mockolate.Tests/ItTests.SequenceEqualsTests.cs
@@ -182,6 +182,128 @@ public sealed partial class ItTests
 			await That(result).IsTrue();
 		}
 
+		public sealed class DoTests
+		{
+			[Fact]
+			public async Task Do_RegistersCallbackForArray()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				int[]? captured = null;
+				((IParameterWithCallback<int[]>)sut).Do(value => captured = value);
+
+				int[] source = [1, 2, 3,];
+				((IParameterMatch<int[]>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForICollection()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				ICollection<int>? captured = null;
+				((IParameterWithCallback<ICollection<int>>)sut).Do(value => captured = value);
+
+				List<int> source = [1, 2, 3,];
+				((IParameterMatch<ICollection<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForIEnumerable()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				IEnumerable<int>? captured = null;
+				((IParameterWithCallback<IEnumerable<int>>)sut).Do(value => captured = value);
+
+				List<int> source = [1, 2, 3,];
+				((IParameterMatch<IEnumerable<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForIList()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				IList<int>? captured = null;
+				((IParameterWithCallback<IList<int>>)sut).Do(value => captured = value);
+
+				List<int> source = [1, 2, 3,];
+				((IParameterMatch<IList<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForIReadOnlyCollection()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				IReadOnlyCollection<int>? captured = null;
+				((IParameterWithCallback<IReadOnlyCollection<int>>)sut).Do(value => captured = value);
+
+				List<int> source = [1, 2, 3,];
+				((IParameterMatch<IReadOnlyCollection<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForIReadOnlyList()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				IReadOnlyList<int>? captured = null;
+				((IParameterWithCallback<IReadOnlyList<int>>)sut).Do(value => captured = value);
+
+				List<int> source = [1, 2, 3,];
+				((IParameterMatch<IReadOnlyList<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForList()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				List<int>? captured = null;
+				((IParameterWithCallback<List<int>>)sut).Do(value => captured = value);
+
+				List<int> source = [1, 2, 3,];
+				((IParameterMatch<List<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForQueue()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				Queue<int>? captured = null;
+				((IParameterWithCallback<Queue<int>>)sut).Do(value => captured = value);
+
+				Queue<int> source = new();
+				source.Enqueue(1);
+				((IParameterMatch<Queue<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+
+			[Fact]
+			public async Task Do_RegistersCallbackForStack()
+			{
+				It.ISequenceEqualsParameter<int> sut = It.SequenceEquals(1, 2, 3);
+				Stack<int>? captured = null;
+				((IParameterWithCallback<Stack<int>>)sut).Do(value => captured = value);
+
+				Stack<int> source = new();
+				source.Push(1);
+				((IParameterMatch<Stack<int>>)sut).InvokeCallbacks(source);
+
+				await That(captured).IsSameAs(source);
+			}
+		}
+
 		public interface ISequenceCollectionConsumer
 		{
 			int WithArray(int[] items);
