+++ Tests/Mockolate.Tests/Web/ItExtensionsTests.IsHttpContentTests.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
@@ -13,6 +14,69 @@ public sealed partial class ItExtensionsTests
 {
 	public sealed partial class IsHttpContentTests
 	{
+		[Fact]
+		public async Task NonGeneric_DispatchesContentThroughCallback()
+		{
+			ItExtensions.IHttpContentParameter sut = It.IsHttpContent();
+			HttpContent? captured = null;
+			sut.Do(content => captured = content);
+
+			MyHttpContent target = new();
+			sut.InvokeCallbacks(target);
+
+			await That(captured).IsSameAs(target);
+		}
+
+		[Fact]
+		public async Task NonGeneric_DispatchesNullThroughCallback()
+		{
+			ItExtensions.IHttpContentParameter sut = It.IsHttpContent();
+			int invocations = 0;
+			HttpContent? captured = new MyHttpContent();
+			sut.Do(content =>
+			{
+				invocations++;
+				captured = content;
+			});
+
+			sut.InvokeCallbacks(null);
+
+			await That(invocations).IsEqualTo(1);
+			await That(captured).IsNull();
+		}
+
+		[Fact]
+		public async Task NonGeneric_IgnoresUnrelatedTypes()
+		{
+			ItExtensions.IHttpContentParameter sut = It.IsHttpContent();
+			int invocations = 0;
+			sut.Do(_ => invocations++);
+
+			sut.InvokeCallbacks("not http content");
+
+			await That(invocations).IsEqualTo(0);
+		}
+
+		[Fact]
+		public async Task NonGenericMatches_ReturnsFalseForNullValue()
+		{
+			ItExtensions.IHttpContentParameter sut = It.IsHttpContent();
+
+			bool result = sut.Matches(null);
+
+			await That(result).IsFalse();
+		}
+
+		[Fact]
+		public async Task NonGenericMatches_ReturnsFalseForUnrelatedType()
+		{
+			ItExtensions.IHttpContentParameter sut = It.IsHttpContent();
+
+			bool result = sut.Matches("not http content");
+
+			await That(result).IsFalse();
+		}
+
 		[Fact]
 		public async Task ShouldSupportMonitoring()
 		{
@@ -105,6 +169,16 @@ public sealed partial class ItExtensionsTests
 				.IsEqualTo(expectSuccess ? HttpStatusCode.OK : HttpStatusCode.NotImplemented);
 		}
 
+		[Fact]
+		public async Task TypedMatches_ReturnsFalseForNullValue()
+		{
+			ItExtensions.IHttpContentParameter sut = It.IsHttpContent();
+
+			bool result = ((IParameterMatch<HttpContent?>)sut).Matches(null);
+
+			await That(result).IsFalse();
+		}
+
 		[Theory]
 		[InlineData("image/png", true)]
 		[InlineData("text/plain", false)]
@@ -124,6 +198,18 @@ public sealed partial class ItExtensionsTests
 
 			await That(result.StatusCode)
 				.IsEqualTo(expectSuccess ? HttpStatusCode.OK : HttpStatusCode.NotImplemented);
+		}
+
+		private sealed class MyHttpContent : HttpContent
+		{
+			protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context)
+				=> Task.CompletedTask;
+
+			protected override bool TryComputeLength(out long length)
+			{
+				length = 0;
+				return false;
+			}
 		}
 	}
 }
